Split ConfigureRoomView into multiple files
